+++ chevah/compat/tests/normal/test_filesystem.py
@@ -16,6 +16,7 @@ from chevah.compat.testing import CompatTestCase, conditionals, manufacture
 
 class FilesystemTestCase(CompatTestCase):
     """
+    Common code for filesystem tests.
     """
 
     def makeLink(self, segments, cleanup=True):
@@ -129,6 +130,17 @@ class TestDefaultFilesystem(FilesystemTestCase):
         self.filesystem.deleteFile(segments)
 
         self.assertFalse(self.filesystem.exists(segments))
+
+    def test_deleteFile_not_found(self):
+        """
+        Return OSError with errno.ENOENT.
+        """
+        segments = [manufacture.string()]
+
+        with self.assertRaises(OSError) as context:
+            self.filesystem.deleteFile(segments)
+
+        self.assertEqual(errno.ENOENT, context.exception.errno)
 
     @conditionals.onCapability('symbolic_link', True)
     def test_deleteFile_file_link(self):